1. Добавлен модуль :features:cash_account
