+++ app/src/main/java/com/example/yandex_school_app/di/module/DomainModule.kt
@@ -3,8 +3,8 @@ package com.example.yandex_school_app.di.module
 import com.example.common.data.repository.TransactionRepositoryImpl
 import com.example.common.domain.repository.TransactionRepository
 import com.example.common.di.ApplicationScope
-import com.example.yandex_school_app.features.cash_account.data.repository.AccountRepositoryImpl
-import com.example.yandex_school_app.features.cash_account.domain.repository.AccountRepository
+import com.example.cash_account.data.repository.AccountRepositoryImpl
+import com.example.cash_account.domain.repository.AccountRepository
 import com.example.category.data.repository.CategoryRepositoryImpl
 import com.example.category.domain.repository.CategoryRepository
 import dagger.Binds
