--- app/src/main/java/com/example/yandex_school_app/features/cash_account/data/datasource/remote/AccountRemoteDataSource.kt
@@ -1,9 +1,0 @@
-package com.example.yandex_school_app.features.cash_account.data.datasource.remote
-
-import com.example.common.data.network.ResponseTemplate
-import com.example.common.domain.entity.AccountDomain
-
-interface AccountRemoteDataSource {
-    suspend fun getAllCashAccount(): ResponseTemplate<List<AccountDomain>>
-    suspend fun createAccount(accountDomain: AccountDomain): ResponseTemplate<AccountDomain>
-}
