--- app/src/main/java/com/example/yandex_school_app/features/cash_account/data/network/model/request/AccountRequestNetwork.kt
@@ -1,7 +1,0 @@
-package com.example.yandex_school_app.features.cash_account.data.network.model.request
-
-data class AccountRequestNetwork(
-    val name: String,
-    val balance: String,
-    val currency: String
-)
