--- app/src/main/java/com/example/yandex_school_app/features/cash_account/data/repository/AccountRepositoryImpl.kt
@@ -1,19 +1,0 @@
-package com.example.yandex_school_app.features.cash_account.data.repository
-
-import com.example.common.data.network.ResponseTemplate
-import com.example.yandex_school_app.features.cash_account.data.datasource.remote.AccountRemoteDataSource
-import com.example.common.domain.entity.AccountDomain
-import com.example.yandex_school_app.features.cash_account.domain.repository.AccountRepository
-import javax.inject.Inject
-
-class AccountRepositoryImpl @Inject constructor(
-    private val accountRemoteDataSource: AccountRemoteDataSource
-) : AccountRepository {
-    override suspend fun getAllCashAccount(): ResponseTemplate<List<AccountDomain>> {
-        return accountRemoteDataSource.getAllCashAccount()
-    }
-
-    override suspend fun createAccount(accountDomain: AccountDomain): ResponseTemplate<AccountDomain> {
-        return accountRemoteDataSource.createAccount(accountDomain)
-    }
-}
