--- app/src/main/java/com/example/yandex_school_app/features/cash_account/domain/entity/Currency.kt
@@ -1,7 +1,0 @@
-package com.example.yandex_school_app.features.cash_account.domain.entity
-
-object Currency {
-    const val RUB = "RUB"
-    const val USD = "USD"
-    const val EUR = "EUR"
-}
