+++ app/src/main/java/com/example/yandex_school_app/features/cash_account/presentation/AccountViewModel.kt
@@ -6,7 +6,7 @@ import com.example.common.data.network.ResponseTemplate
 import com.example.common.presentation.ToastController
 import com.example.common.manager.AccountManager
 import com.example.common.domain.entity.AccountDomain
-import com.example.yandex_school_app.features.cash_account.domain.usecase.AccountUseCase
+import com.example.cash_account.domain.usecase.AccountUseCase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
