+++ app/src/main/java/com/example/yandex_school_app/features/cash_account/presentation/ui/CreateCashAccount.kt
@@ -25,7 +25,7 @@ import androidx.navigation.NavController
 import com.example.yandex_school_app.MainActivity
 import com.example.common.presentation.ToastController
 import com.example.common.domain.entity.AccountDomain
-import com.example.yandex_school_app.features.cash_account.domain.entity.Currency
+import com.example.cash_account.domain.entity.Currency
 import com.example.yandex_school_app.features.cash_account.presentation.AccountViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
