+++ features/cash_account/.gitignore
@@ -1,0 +1,1 @@
+/build
