+++ features/cash_account/build.gradle.kts
@@ -1,23 +1,19 @@
 plugins {
-    alias(libs.plugins.android.application)
+    alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.kotlin.compose)
 
     alias(libs.plugins.ksp)
 }
 
 android {
-    namespace = "com.example.yandex_school_app"
+    namespace = "com.example.cash_account"
     compileSdk = 35
 
     defaultConfig {
-        applicationId = "com.example.yandex_school_app"
         minSdk = 26
-        targetSdk = 35
-        versionCode = 1
-        versionName = "1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles("consumer-rules.pro")
     }
 
     buildTypes {
@@ -36,20 +32,12 @@ android {
     kotlinOptions {
         jvmTarget = "11"
     }
-    buildFeatures {
-        compose = true
-    }
 }
 
 dependencies {
-    implementation(project(":navigation"))
     implementation(project(":common"))
-    implementation(project(":features:settings"))
-    implementation(project(":features:category"))
-    implementation(project(":features:cash_account"))
 
     implementation(libs.androidx.core.ktx)
-    implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
     implementation(platform(libs.androidx.compose.bom))
     implementation(libs.androidx.ui)
@@ -57,19 +45,12 @@ dependencies {
     implementation(libs.androidx.ui.tooling.preview)
     implementation(libs.androidx.material3)
 
-    implementation(libs.airbnb.android.lottie.compose)
-    implementation(libs.androidx.navigation.compose)
     implementation(libs.google.dagger.core)
     ksp(libs.google.dagger.compiler)
     implementation(libs.squareup.retrofit2)
     implementation(libs.squareup.retrofit2.gson)
-    implementation(libs.androidx.lifecycle.viewmodel)
 
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
-    androidTestImplementation(platform(libs.androidx.compose.bom))
-    androidTestImplementation(libs.androidx.ui.test.junit4)
-    debugImplementation(libs.androidx.ui.tooling)
-    debugImplementation(libs.androidx.ui.test.manifest)
 }
