+++ features/cash_account/src/main/java/com/example/cash_account/data/mapper/AccountMapper.kt
@@ -1,0 +1,21 @@
+package com.example.cash_account.data.mapper
+
+import com.example.common.domain.entity.AccountDomain
+import com.example.cash_account.data.network.model.request.AccountRequestNetwork
+import com.example.cash_account.data.network.model.response.AccountResponseNetwork
+import javax.inject.Inject
+
+class AccountMapper @Inject constructor() {
+    fun toAccountDomain(accountNetwork: AccountResponseNetwork) = AccountDomain(
+        id = accountNetwork.id.toInt(),
+        name = accountNetwork.name,
+        balance = accountNetwork.balance,
+        currency = accountNetwork.currency
+    )
+
+    fun toAccountRequest(accountDomain: AccountDomain) = AccountRequestNetwork(
+        name = accountDomain.name,
+        balance = accountDomain.balance,
+        currency = accountDomain.currency
+    )
+}
