+++ features/cash_account/src/main/java/com/example/cash_account/data/network/model/request/AccountRequestNetwork.kt
@@ -1,0 +1,7 @@
+package com.example.cash_account.data.network.model.request
+
+data class AccountRequestNetwork(
+    val name: String,
+    val balance: String,
+    val currency: String
+)
