+++ features/cash_account/src/main/java/com/example/cash_account/data/network/model/response/AccountResponseNetwork.kt
@@ -1,0 +1,11 @@
+package com.example.cash_account.data.network.model.response
+
+data class AccountResponseNetwork(
+    val id: Long,
+    val userId: Long,
+    val name: String,
+    val balance: String,
+    val currency: String,
+    val createdAt: String,
+    val updatedAt: String
+)
