+++ features/cash_account/src/main/java/com/example/cash_account/data/repository/AccountRepositoryImpl.kt
@@ -1,0 +1,19 @@
+package com.example.cash_account.data.repository
+
+import com.example.cash_account.data.datasource.remote.AccountRemoteDataSource
+import com.example.common.data.network.ResponseTemplate
+import com.example.common.domain.entity.AccountDomain
+import com.example.cash_account.domain.repository.AccountRepository
+import javax.inject.Inject
+
+class AccountRepositoryImpl @Inject constructor(
+    private val accountRemoteDataSource: AccountRemoteDataSource
+) : AccountRepository {
+    override suspend fun getAllCashAccount(): ResponseTemplate<List<AccountDomain>> {
+        return accountRemoteDataSource.getAllCashAccount()
+    }
+
+    override suspend fun createAccount(accountDomain: AccountDomain): ResponseTemplate<AccountDomain> {
+        return accountRemoteDataSource.createAccount(accountDomain)
+    }
+}
