+++ features/cash_account/src/main/java/com/example/cash_account/domain/entity/Currency.kt
@@ -1,0 +1,7 @@
+package com.example.cash_account.domain.entity
+
+object Currency {
+    const val RUB = "RUB"
+    const val USD = "USD"
+    const val EUR = "EUR"
+}
