+++ features/cash_account/src/main/java/com/example/cash_account/domain/usecase/AccountUseCase.kt
@@ -1,0 +1,13 @@
+package com.example.cash_account.domain.usecase
+
+import com.example.common.domain.entity.AccountDomain
+import com.example.cash_account.domain.repository.AccountRepository
+import javax.inject.Inject
+
+class AccountUseCase @Inject constructor(
+    private val accountRepository: AccountRepository
+) {
+    suspend fun getAllCashAccount() = accountRepository.getAllCashAccount()
+    suspend fun createCashAccount(accountDomain: AccountDomain) =
+        accountRepository.createAccount(accountDomain)
+}
