+++ settings.gradle.kts
@@ -26,3 +26,4 @@ include(":features")
 include(":common")
 include(":features:settings")
 include(":features:category")
+include(":features:cash_account")
